app/requirements.txt
accelerate==0.23.0
bitsandbytes==0.41.1
cachetools==5.5.0
faiss-cpu==1.9.0
fastapi==0.115.0
fastapi-limiter==0.1.6
gradio==3.48.0
hypercorn==0.17.3
openai==1.60.2
protobuf==3.20.3
pydantic==2.10.2
pydantic-core==2.27.1
scipy==1.10.1
sentence-transformers==3.4.1
sentencepiece==0.1.99
spaces==0.16.1
sympy==1.12
tokenizers>=0.14.0
torch>=2.0
transformers>=4.35.0
uvloop==0.21.0
